Checks for permissions when creating/viewing/updating/deleting, remove privilege mostly implemented and tests mostly updated

+++ src/Service/Modules/PermissionsModule.cs
@@ -1,17 +1,28 @@
 ﻿namespace Linn.Authorisation.Service.Modules
 {
+    using System;
+    using System.Linq;
+
     using Linn.Authorisation.Facade;
     using Linn.Authorisation.Resources;
+    using Linn.Common.Authorisation;
+    using Linn.Common.Facade;
+    using Linn.Production.Domain.LinnApps;
+    using Linn.Production.Service.Extensions;
+
     using Nancy;
     using Nancy.ModelBinding;
+    using Nancy.Security;
 
     public sealed class PermissionsModule : NancyModule
     {
         private readonly IPermissionService permissionService;
+        private readonly IAuthorisationService authorisationService;
 
-        public PermissionsModule(IPermissionService permissionService)
+        public PermissionsModule(IPermissionService permissionService, IAuthorisationService authorisationService)
         {
             this.permissionService = permissionService;
+            this.authorisationService = authorisationService;
             this.Post("/authorisation/permissions", _ => this.CreatePermission());
             this.Delete("/authorisation/permissions", _ => this.RemovePermission());
             this.Get("/authorisation/permissions/{id:int}", parameters => this.GetPermissionsForPrivilege(parameters.id));
@@ -19,6 +30,14 @@
 
         private object CreatePermission()
         {
+            this.RequiresAuthentication();
+            var privileges = this.Context?.CurrentUser?.GetPrivileges().ToList();
+
+            if (!this.authorisationService.HasPermissionFor(AuthorisedAction.AuthorisationAdmin, privileges))
+            {
+                return this.Negotiate.WithModel((new BadRequestResult<string>("You are not authorised to create permissions")));
+            }
+
             var resource = this.Bind<PermissionResource>();
             var result = this.permissionService.CreatePermission(resource);
             return this.Negotiate.WithModel(result);
@@ -26,6 +45,14 @@
 
         private object RemovePermission()
         {
+            this.RequiresAuthentication();
+            var privileges = this.Context?.CurrentUser?.GetPrivileges().ToList();
+
+            if (!this.authorisationService.HasPermissionFor(AuthorisedAction.AuthorisationAdmin, privileges))
+            {
+                return this.Negotiate.WithModel((new BadRequestResult<string>("You are not authorised to remove permissions")));
+            }
+
             var resource = this.Bind<PermissionResource>();
             var result = this.permissionService.RemovePermission(resource);
             return this.Negotiate.WithModel(result);
@@ -33,6 +60,14 @@
 
         private object GetPermissionsForPrivilege(int privilegeId)
         {
+            this.RequiresAuthentication();
+            var privileges = this.Context?.CurrentUser?.GetPrivileges().ToList();
+
+            if (!this.authorisationService.HasPermissionFor(AuthorisedAction.AuthorisationAdmin, privileges))
+            {
+                return this.Negotiate.WithModel((new BadRequestResult<string>("You are not authorised to view permissions")));
+            }
+
             var resource = this.Bind<PermissionResource>();
             var result = this.permissionService.GetAllPermissionsForPrivilege(privilegeId);
             return this.Negotiate.WithModel(result);
